refactor: rename redux folder to store
