+++ pages/_app.tsx
@@ -1,6 +1,6 @@
 import React, { FC } from 'react'
 import { AppProps } from 'next/app'
-import { wrapper } from 'src/redux'
+import { wrapper } from 'src/store'
 import Layout from 'src/components/Layout'
 import GlobalBaseStyle from 'src/styles/GlobalBaseStyle'
