+++ src/components/User/index.tsx
@@ -1,6 +1,6 @@
 import React, { FC, useEffect } from 'react'
 import { useSelector, useDispatch } from 'react-redux'
-import { getProfileRequest } from 'src/redux/user/actions'
+import { getProfileRequest } from 'src/store/user/actions'
 import Profile from './Profile'
 
 const User: FC = () => {
