+++ src/store/rootSagas.ts
@@ -1,7 +1,7 @@
 import { all, takeLatest, AllEffect, ForkEffect } from 'redux-saga/effects'
 
-import { actionTypes } from 'src/redux/user/actions'
-import { updateProfileRequest } from 'src/redux/user/saga'
+import { actionTypes } from 'src/store/user/actions'
+import { updateProfileRequest } from 'src/store/user/saga'
 
 function* rootSaga(): Generator<AllEffect<ForkEffect<never>>, void, unknown> {
   yield all([takeLatest(actionTypes.GET_PROFILE_REQUEST, updateProfileRequest)])
